Add job spawn route system to frontend, grouped by labels
This commit makes it possible to spawn a PtpJob from the frontend via the `send-ptp-job` route, launching the Python scripts per label. The `ptp.py` is modified accordingly.
Styling is also changed, to reflect this grouping and to allow the container to show the annotations well. A temporary directory system is introduced in the config.