Refactor Point to Polygon Pipeline
This commit introduces several fixes and changes regarding the Point to Polygon pipeline. Essentially, there is now just one pipeline, triggered by the selected button, that first computes the expected area, then uploads it on the DB. The expected area is then use to apply the Point to Polygon conversion. In the next commits a Job that uploads the new annotations will be created, as well as a better system for the exchange of information between jobs.