Unify Ptp jobs into one, make the Ptp job use a single python script call
This commit changes the Ptp job, so that it will execute all the existing part (compute expected area and upload the converted polygons to the DB) in a single Job. Also, there will not be a need to upload the expected area to the DB, as it is done within the python script. This makes the script more effective, as explained by comments within the PR.