Fix handling failure of generating a job. Add test for this case.
This commit fixes the case of when the application fails to dispatch a PTP job.
If the job fails to be dispatched/generated, it will now throw a 500 error and ask the user to try later. At the same time, the internal logger will warn us of the error.
This commit also adds a test for this case to the existing ones.